# feat: add ASCII fast path optimization to ShapeWord::build()