related to #308

Skip expensive grapheme iteration for simple ASCII words
which helps with code, ASCII, and mixed Unicode content.

In my benchmarks

**Test Environment**
Hardware: Linux system
Methodology: 10 runs per test, median time reported
Warm-up: 3 iterations before measurement
Test Scale: 100K-200K characters per test case


| Test Case                 | Original (chars/ms) | Optimized (chars/ms) | **Improvement** |
|---------------------------|--------------------:|-----------------:|----------------:|
| Many Short ASCII Words   |              1,400.6 |            1,492.3 |      **+6.6%** |
| Code-like Text            |            283,946.9 |          375,810.4 |     **+32.3%** |
| English Prose             |                985.4 |            1,052.6 |      **+6.8%** |
| Mixed ASCII/Unicode       |                914.9 |              954.3 |      **+4.3%** |
| Complex Unicode           |                648.6 |              654.1 |      **+0.8%** |
| Sample File (100K chars)  |             17,664.7 |           18,515.1 |      **+4.8%** |

stupid me created a new micro bench for myself and have not used yours.

| Test Name | Time | Performance Change |
|-----------|------|-------------------|
| **Wrap(None, Simple)/small amount of text** | 1.7443 µs | **-14.70%** ⬆️ |
| **Wrap(None, Simple)/large amount of text** | 2.0029 ms | **-11.35%** ⬆️ |
| **Wrap(None, Simple)/arabic text** | 85.286 µs | **-2.75%** ⬆️ |
| **Wrap(None, Simple)/hebrew text** | 95.415 µs | **-6.50%** ⬆️ |
| **Wrap(None, Simple)/emoji text** | 55.869 ms | **-10.23%** ⬆️ |
| **Wrap(None, Advanced)/small amount of text** | 9.5349 µs | **-2.98%** ⬆️ |
| **Wrap(None, Advanced)/large amount of text** | 13.257 ms | **-2.85%** ⬆️ |
| **Wrap(None, Advanced)/arabic text** | 1.0288 ms | **-2.20%** ⬆️ |
| **Wrap(None, Advanced)/hebrew text** | 580.61 µs | **-1.64%** ⬆️ |
| **Wrap(None, Advanced)/emoji text** | 333.86 ms | **-2.90%** ⬆️ |
| **Wrap(Glyph, Simple)/small amount of text** | 1.7443 µs | **-14.70%** ⬆️ |
| **Wrap(Glyph, Simple)/large amount of text** | 1.9955 ms | **-11.61%** ⬆️ |
| **Wrap(Glyph, Simple)/arabic text** | 85.666 µs | **-1.37%** ⬆️ |
| **Wrap(Glyph, Simple)/hebrew text** | 96.295 µs | **-5.30%** ⬆️ |
| **Wrap(Glyph, Simple)/emoji text** | 56.283 ms | **-11.07%** ⬆️ |
| **Wrap(Glyph, Advanced)/small amount of text** | 9.4349 µs | **-4.42%** ⬆️ |
| **Wrap(Glyph, Advanced)/large amount of text** | 13.294 ms | **-3.97%** ⬆️ |
| **Wrap(Glyph, Advanced)/arabic text** | 1.0342 ms | **-1.39%** ⬆️ |
| **Wrap(Glyph, Advanced)/hebrew text** | 580.45 µs | No change |
| **Wrap(Glyph, Advanced)/emoji text** | 333.07 ms | **-3.29%** ⬆️ |
| **Wrap(Word, Simple)/small amount of text** | 1.7443 µs | **-14.70%** ⬆️ |
| **Wrap(Word, Simple)/large amount of text** | 2.0029 ms | **-11.35%** ⬆️ |
| **Wrap(Word, Simple)/arabic text** | 85.286 µs | **-2.75%** ⬆️ |
| **Wrap(Word, Simple)/hebrew text** | 95.415 µs | **-6.50%** ⬆️ |
| **Wrap(Word, Simple)/emoji text** | 55.869 ms | **-10.23%** ⬆️ |
| **Wrap(Word, Advanced)/small amount of text** | 9.5349 µs | **-2.98%** ⬆️ |
| **Wrap(Word, Advanced)/large amount of text** | 13.257 ms | **-2.85%** ⬆️ |
| **Wrap(Word, Advanced)/arabic text** | 1.0288 ms | **-2.20%** ⬆️ |
| **Wrap(Word, Advanced)/hebrew text** | 580.61 µs | **-1.64%** ⬆️ |
| **Wrap(Word, Advanced)/emoji text** | 333.86 ms | **-2.90%** ⬆️ |
| **load FontSystem** | 6.1215 ms | +1.27% |